### 2025-12-29 #9702 @editorialbot
[PRE REVIEW]: JupyterGIS: A Collaborative GIS Environment for JupyterLab
**Submitting author:** @SylvainCorlay (<a href="http://orcid.org/0009-0007-2816-4102">Sylvain Corlay</a>)
**Repository:** https://github.com/geojupyter/jupytergis/
**Branch with paper.md** (empty if default branch): main
**Version:** 0.11.1
