### 2025-12-29 eef783e
Add GitHub aliases to each author
