+++ CITATION.cff
@@ -31,10 +31,12 @@ authors:
   - name: "The GeoJupyter Community"
   - family-names: "Renou"
     given-names: "Martin"
+    alias: "martinRenou"
     orcid: "https://orcid.org/0009-0002-3422-9156"
     affiliation: "QuantStack, France"
   - family-names: "Verma"
     given-names: "Arjun"
+    alias: "arjxn-py"
     orcid: "https://orcid.org/0009-0007-3278-9607"
     affiliation: "QuantStack, France"
   - family-names: "Fisher"
@@ -43,22 +45,27 @@ authors:
     affiliation: "Eric and Wendy Schmidt Center for Data Science & Environment at UC Berkeley, United States"
   - family-names: "Mooney"
     given-names:  "Gregory"
+    alias: "gjmooney"
     orcid: "https://orcid.org/0009-0007-5501-6471"
     affiliation: "QuantStack, France"
   - family-names: "Brichet"
     given-names: "Nicolas"
+    alias: "brichet"
     orcid: "https://orcid.org/0009-0009-2649-1809"
     affiliation: "QuantStack, France"
   - family-names: "Brochart"
     given-names: "David"
+    alias: "davidbrochart"
     orcid: "https://orcid.org/0009-0007-2318-0285"
     affiliation: "QuantStack, France"
   - family-names: "Fouilloux"
     given-names: "Anne"
+    alias: "annefou"
     orcid: "https://orcid.org/0000-0002-1784-2920"
     affiliation: "LifeWatch ERIC, Spain"
   - family-names: "Corlay"
     given-names: "Sylvain"
+    alias: "SylvainCorlay"
     orcid: "https://orcid.org/0009-0007-2816-4102"
     affiliation: "QuantStack, France"
 
