### 2025-12-29 43688d8
Recognize Greg contributed more than Matt
Based on number of contributions pulled from GitHub API.
After top 5, will sort alphabetically.
I don't know if this is the right decision, this is just based on me
feeling weird about being listed before Greg, and compromising between
recognizing volume of contributions with maintainability of the
author list.
