+++ CITATION.cff
@@ -39,15 +39,17 @@ authors:
     alias: "arjxn-py"
     orcid: "https://orcid.org/0009-0007-3278-9607"
     affiliation: "QuantStack, France"
-  - family-names: "Fisher"
-    given-names: "Matt"
-    orcid: "https://orcid.org/0000-0003-3260-5445"
-    affiliation: "Eric and Wendy Schmidt Center for Data Science & Environment at UC Berkeley, United States"
   - family-names: "Mooney"
     given-names:  "Gregory"
     alias: "gjmooney"
     orcid: "https://orcid.org/0009-0007-5501-6471"
     affiliation: "QuantStack, France"
+  - family-names: "Fisher"
+    given-names: "Matt"
+    alias: "mfisher87"
+    website: "https://mfisher87.github.io/"
+    orcid: "https://orcid.org/0000-0003-3260-5445"
+    affiliation: "Eric and Wendy Schmidt Center for Data Science & Environment at UC Berkeley, United States"
   - family-names: "Brichet"
     given-names: "Nicolas"
     alias: "brichet"
