### 2025-12-29 2d947d0
Add all contributors from GitHub contributors API, sort alphabetically
Using script: <https://github.com/mfisher87/cff-authors-from-gh-contributors>
