+++ CITATION.cff
@@ -29,6 +29,8 @@ date-released: "2025-12-18"
 # TODO: More comprehensive authors list. Initialized from the JOSS paper.
 authors:
   - name: "The GeoJupyter Community"
+
+  # Top 5, sorted by # of contributions returned by the GitHub API
   - family-names: "Renou"
     given-names: "Martin"
     alias: "martinRenou"
@@ -55,21 +57,94 @@ authors:
     alias: "brichet"
     orcid: "https://orcid.org/0009-0009-2649-1809"
     affiliation: "QuantStack, France"
+
+  # All contributors sorted alphabetically
+  - given-names: "Edward"
+    family-names: "Betts"
+    alias: "EdwardBetts"
+    website: "http://edwardbetts.com/"
+  - given-names: "Paul"
+    family-names: "Blottiere"
+    alias: "pblottiere"
+    website: "https://pblottiere.github.io/"
   - family-names: "Brochart"
     given-names: "David"
     alias: "davidbrochart"
     orcid: "https://orcid.org/0009-0007-2318-0285"
     affiliation: "QuantStack, France"
-  - family-names: "Fouilloux"
-    given-names: "Anne"
-    alias: "annefou"
-    orcid: "https://orcid.org/0000-0002-1784-2920"
-    affiliation: "LifeWatch ERIC, Spain"
+  - given-names: "Jessenia"
+    family-names: "Chan"
+    alias: "JesseniaChan"
   - family-names: "Corlay"
     given-names: "Sylvain"
     alias: "SylvainCorlay"
     orcid: "https://orcid.org/0009-0007-2816-4102"
     affiliation: "QuantStack, France"
+  - given-names: "Kristin"
+    family-names: "Davis"
+    alias: "kpdavi"
+  - given-names: "Tyler"
+    family-names: "Erickson"
+    alias: "tylere"
+    website: "https://www.analyze.earth/"
+  - given-names: "Matthew"
+    family-names: "Feickert"
+    alias: "matthewfeickert"
+    website: "http://www.matthewfeickert.com/"
+  - given-names: "Elifsu"
+    family-names: "Filiz"
+    alias: "elifsu-simula"
+  - family-names: "Fouilloux"
+    given-names: "Anne"
+    alias: "annefou"
+    orcid: "https://orcid.org/0000-0002-1784-2920"
+    affiliation: "LifeWatch ERIC, Spain"
+  - given-names: "Annie"
+    family-names: "Gao"
+    alias: "gaolinai8888"
+  - given-names: "Jason"
+    family-names: "Grout"
+    alias: "jasongrout"
+    website: "https://jasongrout.org"
+  - given-names: "Florence"
+    family-names: "Haudin"
+    alias: "HaudinFlorence"
+  - given-names: "Meriem Ben"
+    family-names: "Ismail"
+    alias: "Meriem-BenIsmail"
+  - given-names: "Duc Trung"
+    family-names: "Le"
+    alias: "trungleduc"
+  - given-names: "Stace"
+    family-names: "Maples"
+    alias: "mapninja"
+    website: "https://gis.stanford.edu"
+  - given-names: "Matthias"
+    family-names: "Meschede"
+    alias: "MMesch"
+    website: "https://mmesch.com"
+  - given-names: "Konstantin"
+    family-names: "Taletskiy"
+    alias: "ktaletsk"
+    website: "https://taletskiy.com"
+  - given-names: "Max"
+    family-names: "Taniguchi-King"
+    alias: "mtaniguchiking"
+  - given-names: "Nakul"
+    family-names: "Verma"
+    alias: "nakul-py"
+  - given-names: "Qiusheng"
+    family-names: "Wu"
[…]
