+++ CITATION.cff
@@ -150,7 +150,7 @@ authors:
 # TODO: Fill in grant numbers. Are there more grants?
 # references:
 #   - type: "grant"
-#     intitution:
+#     institution:
 #       name: "European Space Agency"
 #     number: "TODO"
 #   - type: "grant"
