### 2026-01-02 d43ffe3
Bump the gha-dependencies group across 1 directory with 5 updates (#1052)
Bumps the gha-dependencies group with 5 updates in the / directory:

| Package | From | To |
| --- | --- | --- |
| [actions/checkout](https://github.com/actions/checkout) | `4` | `6` |
| [actions/cache](https://github.com/actions/cache) | `4` | `5` |
| [actions/upload-artifact](https://github.com/actions/upload-artifact) | `4` | `6` |
| [actions/download-artifact](https://github.com/actions/download-artifact) | `6` | `7` |
| [dawidd6/action-download-artifact](https://github.com/dawidd6/action-download-artifact) | `11` | 
[…]
