### 2026-01-05 19cff8d
Add Story mode presentation toggle button to toolbar
