+++ packages/base/src/commands/BaseCommandIDs.ts
@@ -68,3 +68,4 @@ export const showIdentifyPanelTab = 'jupytergis:showIdentifyPanelTab';
 
 // Story maps
 export const addStorySegment = 'jupytergis:addStorySegment';
+export const toggleStoryPresentationMode = 'jupytergis:toggleStoryPresentationMode'
