### 2026-01-05 990e4e3
Add Story mode presentation toggle button to toolbar
