+++ packages/base/src/commands/BaseCommandIDs.ts
@@ -68,4 +68,5 @@ export const showIdentifyPanelTab = 'jupytergis:showIdentifyPanelTab';
 
 // Story maps
 export const addStorySegment = 'jupytergis:addStorySegment';
-export const toggleStoryPresentationMode = 'jupytergis:toggleStoryPresentationMode'
+export const toggleStoryPresentationMode =
+  'jupytergis:toggleStoryPresentationMode';
