+++ CITATION.cff
@@ -90,7 +90,7 @@ authors:
   - given-names: "Matthew"
     family-names: "Feickert"
     alias: "matthewfeickert"
-    website: "http://www.matthewfeickert.com/"
+    orcid: "https://orcid.org/0000-0003-4124-7862"
   - given-names: "Elifsu"
     family-names: "Filiz"
     alias: "elifsu-simula"
