### 2026-01-05 7a89bb9
[pre-commit.ci] pre-commit autoupdate (#1057)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
