+++ packages/schema/src/model.ts
@@ -334,6 +334,7 @@ export class JupyterGISModel implements IJupyterGISModel {
       this.sharedModel.sources = jsonData.sources ?? {};
       this.sharedModel.layers = jsonData.layers ?? {};
       this.sharedModel.layerTree = jsonData.layerTree ?? [];
+      this.sharedModel.stories = jsonData.stories ?? {};
       this.sharedModel.options = jsonData.options ?? {
         latitude: 0,
         longitude: 0,
