### 2026-01-07 9f5ae9a
Hide top bars
