### 2026-01-08 12b77cc
Disable commands in specta
