### 2026-01-08 3687c29
Only enable specta presentation mode when stories are present
