### 2026-01-08 87a62c7
Only enable specta presentation mode when stories are present
