+++ python/jupytergis_core/src/factory.ts
@@ -45,8 +45,6 @@ export class JupyterGISDocumentWidgetFactory extends ABCWidgetFactory<
     this._commands = options.commands;
     this._externalCommandRegistry = externalCommandRegistry;
     this._contentsManager = options.manager?.contents;
-
-    this._isSpecta = window.location.pathname.includes('specta');
   }
 
   /**
@@ -91,8 +89,6 @@ export class JupyterGISDocumentWidgetFactory extends ABCWidgetFactory<
       externalCommands: this._externalCommandRegistry.getCommands(),
     });
     
-    toolbar.setHidden(this._isSpecta);
-    
     return new JupyterGISDocumentWidget({
       context,
       content,
@@ -100,7 +96,6 @@ export class JupyterGISDocumentWidgetFactory extends ABCWidgetFactory<
     });
   }
 
-  private _isSpecta: boolean
   private _commands: CommandRegistry;
   private _externalCommandRegistry: IJGISExternalCommandRegistry;
   private _backendCheck?: () => boolean;
