+++ packages/schema/src/model.ts
@@ -606,6 +606,19 @@ export class JupyterGISModel implements IJupyterGISModel {
   }
 
   /**
+   * Check if the application is running in Specta mode.
+   * Specta mode is enabled when the URL contains 'specta' AND the model has stories.
+   *
+   * @returns True if running in Specta mode
+   */
+  isSpectaMode(): boolean {
+    const hasStories = Object.keys(this.sharedModel.stories).length > 0;
+    const urlContainsSpecta = window.location.pathname.includes('specta');
+    
+    return urlContainsSpecta && hasStories;
+  }
+
+  /**
    * Placeholder in case we eventually want to support multiple stories
    * @returns First/only story
    */
