### 2026-01-08 2ffed6b
Show a + after results if there are more results
