+++ packages/base/src/stacBrowser/types/types.ts
@@ -244,6 +244,6 @@ export type StacFilterSetters = Record<
 // Shared type for setResults function signature
 export type SetResultsFunction = (
   results: IStacItem[],
-  totalResults: number,
+  totalResults: string,
   totalPages: number,
 ) => void;
