### 2026-01-08 7d8b627
Update rasterlayer_gallery_generator.py to add Macrostrat Carto raster layer. Update the rasterlayer_gallery directory.
