### 2026-01-08 4b6bc21
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
