### 2026-01-09 2a6b30f
Dont hardcode proxy strategy
