### 2026-01-11 6477f3e
fix(qgis_loader): Allow import using system installation
