### 2026-01-11 c377263
fix(qgis_loader): using standard qgis.PyQt import instead of direct PyQt5
