### 2026-01-11 3cf20e7
add myself to CITATION.cff
