### 2026-01-12 40c45c8
Update specta check logic
