### 2026-01-12 f612ebe
Update specta check logic
