### 2026-01-12 abac47b
No specta for unguided mode
