### 2026-01-12 290ff0c
No specta for unguided mode
