### 2026-01-12 057bae0
Update gallery to support more layer types and sources.
