+++ packages/base/rasterlayer_gallery_generator.py
@@ -1,252 +1,0 @@
-from datetime import date, timedelta
-import json
-from io import BytesIO
-import os
-
-import requests
-from PIL import Image
-import mercantile
-from xyzservices import providers, TileProvider
-
-THUMBNAILS_LOCATION = "rasterlayer_gallery"
-
-
-def fetch_tile(url_template, x, y, z, s="a"):
-    """
-    Fetch a tile from the given URL template.
-    """
-    url = url_template.format(x=x, y=y, z=z, s=s)
-    print(f"   Fetch {url}")
-    response = requests.get(
-        url, headers={"Content-Type": "application/json", "User-Agent": "JupyterGIS"}
-    )
-    response.raise_for_status()
-    return Image.open(BytesIO(response.content))
-
-
-def latlng_to_tile(lat, lng, zoom):
-    """
-    Convert latitude/longitude to tile coordinates.
-    """
-    tile = mercantile.tile(lng, lat, zoom, True)
-    return tile.x, tile.y
-
-
-def create_thumbnail(
-    url_template,
-    lat,
-    lng,
-    zoom,
-    thumbnail_path,
-    tile_size=256,
-    thumbnail_size=(512, 512),
-):
-    """
-    Create a thumbnail for the specified location and zoom level.
-    """
-    # Skip if thumbnail already exists
-    if os.path.exists(thumbnail_path):
-        return Image.open(thumbnail_path)
-    x, y = latlng_to_tile(lat, lng, zoom)
-
-    # Fetch the tiles (2x2 grid for the thumbnail)
-    tiles = []
-    for dy in range(2):
-        row = []
-        for dx in range(2):
-            tile_x, tile_y = x + dx, y + dy
-            tile = fetch_tile(url_template, tile_x, tile_y, zoom)
-            row.append(tile)
-        tiles.append(row)
-
-    # Create a blank image for the thumbnail
-    thumbnail = Image.new("RGB", (2 * tile_size, 2 * tile_size))
-
-    # Paste the tiles into the thumbnail image
-    for dy, row in enumerate(tiles):
-        for dx, tile in enumerate(row):
-            thumbnail.paste(tile, (dx * tile_size, dy * tile_size))
-
-    # Resize to the desired thumbnail size
-    thumbnail = thumbnail.resize(thumbnail_size, Image.LANCZOS)
-    return thumbnail
-
-
-yesterday = (date.today() - timedelta(days=1)).strftime("%Y-%m-%d")
-
-# San Francisco
-san_francisco = {"lat": 37.7749, "lng": -122.4194, "zoom": 5}
-
-middle_europe = {"lat": 48.63290858589535, "lng": -350.068359375, "zoom": 4}
-
-# Default
-france = {"lat": 47.040182144806664, "lng": 1.2963867187500002, "zoom": 5}
-
-thumbnails_providers_positions = {
-    "OpenStreetMap": {
-        "Special Rules": {
-            "BZH": {"lat": 47.76702233051035, "lng": -3.4675598144531254, "zoom": 8},
-            "CH": {"lat": 46.8182, "lng": 8.2275, "zoom": 8},
-            "DE": {"lat": 51.1657, "lng": 10.4515, "zoom": 8},
-            "France": france,
-            "HOT": france,
-        },
-        "Default": france,
-    },
-    "NASAGIBS": {"Special Rules": {}, "Default": france},
-    # 'JusticeMap': {
-    #     'Special Rules': {},
-    #     'Default': san_francisco,
-    # },
-    "USGS": {
-        "Special Rules": {},
-        "Default": san_francisco,
-    },
-    "WaymarkedTrails": {
-        "Special Rules": {},
-        "Default": france,
-    },
[…]
