+++ packages/schema/src/interfaces.ts
@@ -281,7 +281,7 @@ export interface IJupyterGISModel extends DocumentRegistry.IModel {
 
   disposed: ISignal<any, void>;
   getSelectedStory(): {
-    storySegmentId: string;
+    storyId: string;
     story: IJGISStoryMap | undefined;
   };
   addStorySegment(): { storySegmentId: string; storyMapId: string } | null;
