+++ packages/schema/src/interfaces.ts
@@ -284,7 +284,7 @@ export interface IJupyterGISModel extends DocumentRegistry.IModel {
     storyId: string;
     story: IJGISStoryMap | undefined;
   };
-  addStorySegment(): { storySegmentId: string; storyMapId: string } | null;
+  addStorySegment(): { storySegmentId: string; storyId: string } | null;
   isSpectaMode(): boolean;
 }
 
