### 2026-01-12 afbe610
Update rasterlayer_gallery_generator.py to check if a thumbnail already exists for each layer. Add a manually generated thumbnail for MarcroStrat.CartoRaster layer.
