### 2026-01-12 18b2c0f
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
