### 2026-01-12 eb4ab5e
Add MacroStrat tile layer to the layer browser (#1060)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
