### 2026-01-12 aef819e
fixup! fix(qgis_loader): Allow import using system installation
- Append path instead of insert at the bottom
