### 2026-01-13 c04398c
Add contributor documentation about ui testing and dependency security
