+++ docs/contributor_guide/how-tos/code_quality.md
@@ -1,3 +1,7 @@
+---
+title: "How to: check code quality"
+---
+
 # Check code quality
 
 ```{seealso}
@@ -10,6 +14,8 @@ We have several tools configured for checking code quality:
   Install checks with `pre-commit install`.
   Run them manually with `pre-commit run --all-files`.
   **Will exit non-zero when finding errors or changing files.**
+  Alternative: Comment `pre-commit.ci autofix` in a PR.
+  This may not fix all errors, as not all errors are autofixable.
   - Ruff formats and lints (sometimes autofixes) Python code.
 
   - Generic pre-commit checks help avoid common mistakes like committing large
@@ -27,14 +33,6 @@ We have several tools configured for checking code quality:
 - **UI tests using [Galata](https://github.com/jupyterlab/galata)**, defined in the
   `ui-tests/` directory.
 
-  ```{warning}
-  Some UI tests depend on snapshot (screenshot) comparison, which depends strongly on
-  the system used to generate the snapshots. Variables like GPU, operating system,
-  display scaling, and more can affect the outcome of snapshot testing. The instructions
-  below will skip snapshot tests for this reason. Please **do not** run snapshot tests
-  locally and instead rely on CI (GitHub Actions).
-  ```
-
   To run locally:
 
   ```bash
@@ -44,13 +42,4 @@ We have several tools configured for checking code quality:
   jlpm run test:local      # Or, to test in jupyterlite, run `test:locallite`
   ```
 
-  - DOM testing: Interact with the application programmatically and verify
-    resulting DOM looks correct.
-
-  - Snapshot testing: Interact with the application programmatically and verify
-    the resulting rendered content exactly matches snapshots in this
-    repository.
-    When snapshot tests fail, new snapshots can be generated by
-    commenting `please update snapshots` in a PR or locally running
-    `jlpm run test:update` from the `ui-tests/` directory.
-    Please carefully verify the snapshot changes are as expected before merging.
+  For more, see :doc:`/contributor-guide/explanation/ui-testing.md`.

[…]
