### 2026-01-13 fc0c208
Check links job: Ignore links to medium blog posts
The same problem as `blog.jupyter.org` links. They seem to reject
requests that don't come from browsers.
