+++ .github/workflows/build.yml
@@ -268,6 +268,7 @@ jobs:
             https://www\.npmjs\.com.*
             ^/.*
             https://blog\.jupyter\.org/.*
+            https://medium\.com/.*
 
   build-lite:
     name: Build JupyterLite artifacts
