### 2026-01-13 5e40171
Eliminate an unnecessary heading level
