### 2026-01-13 2a002f7
Remove title overrides from crossrefs
