### 2026-01-13 a058f28
Remove redundant titles
I was hoping that specifying a title and h1 title would enable me to
control the text of crossrefs without writing custom crossref titles,
while keeping the document title shown in the sidebar the same. This
doesn't work.
