+++ docs/contributor_guide/how-tos/ui-tests-update-snapshots.md
@@ -1,7 +1,3 @@
----
-title: 'How to: update UI test snapshots'
----
-
 # UI tests: Update snapshots
 
 ## Background
