### 2026-01-13 d3ee0dc
Fix bot permission handling
