### 2026-01-13 1387d7c
Bot: provide negative response if the user has no permission (#1069)
