+++ .github/workflows/update_galata_references.yaml
@@ -35,10 +35,11 @@ jobs:
         if: |
           !contains(fromJSON('["OWNER","COLLABORATOR","MEMBER"]'), steps.association.outputs.association)
         run: |
+          gh api repos/${{ github.repository }}/issues/comments/${{ github.event.comment.id }}/reactions --raw-field 'content=-1'
           echo "User not authorized to update snapshots"
           exit 1
 
-      - name: React to the triggering comment
+      - name: React positively to the triggering comment
         run: |
           gh api repos/${{ github.repository }}/issues/comments/${{ github.event.comment.id }}/reactions --raw-field 'content=+1'
         env:
