+++ .github/workflows/update_galata_references.yaml
@@ -38,6 +38,8 @@ jobs:
           gh api repos/${{ github.repository }}/issues/comments/${{ github.event.comment.id }}/reactions --raw-field 'content=-1'
           echo "User not authorized to update snapshots"
           exit 1
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
       - name: React positively to the triggering comment
         run: |
