### 2026-01-13 f6e5229
Bot: Reusable workflow for checking permissions
