### 2026-01-13 9b8141c
Bot: Reusable workflow for checking permissions (#1072)
