### 2026-01-13 4d2de7e
Update snapshot update how-to to reflect the bot works again
