+++ .github/workflows/update_galata_references.yaml
@@ -20,49 +20,9 @@ jobs:
     if: github.event.issue.pull_request && contains(github.event.comment.body, 'please update snapshots')
 
     steps:
-      - uses: martinRenou/maintainer-tools/.github/actions/update-snapshots-check-permission@check_perm
+      - uses: martinRenou/maintainer-tools/.github/actions/update-snapshots-checkout@check_perm
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Checkout
-        uses: actions/checkout@v6
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Get PR Info
-        id: pr
-        env:
-          PR_NUMBER: ${{ github.event.issue.number }}
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          GH_REPO: ${{ github.repository }}
-          COMMENT_AT: ${{ github.event.comment.created_at }}
-        run: |
-          pr="$(gh api /repos/${GH_REPO}/pulls/${PR_NUMBER})"
-          head_sha="$(echo "$pr" | jq -r .head.sha)"
-          pushed_at="$(echo "$pr" | jq -r .pushed_at)"
-
-          if [[ $(date -d "$pushed_at" +%s) -gt $(date -d "$COMMENT_AT" +%s) ]]; then
-              echo "Updating is not allowed because the PR was pushed to (at $pushed_at) after the triggering comment was issued (at $COMMENT_AT)"
-              exit 1
-          fi
-
-          echo "head_sha=$head_sha" >> $GITHUB_OUTPUT
-
-      - name: Checkout the branch from the PR that triggered the job
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: gh pr checkout ${{ github.event.issue.number }}
-
-      - name: Validate the fetched branch HEAD revision
-        env:
-          EXPECTED_SHA: ${{ steps.pr.outputs.head_sha }}
-        run: |
-          actual_sha="$(git rev-parse HEAD)"
-
-          if [[ "$actual_sha" != "$EXPECTED_SHA" ]]; then
-              echo "The HEAD of the checked out branch ($actual_sha) differs from the HEAD commit available at the time when trigger comment was submitted ($EXPECTED_SHA)"
-              exit 1
-          fi
 
       - name: Download extension package
         uses: dawidd6/action-download-artifact@v12
