### 2026-01-14 52edb16
Bot: Use official jupyterlab/maintainer-tools action
