### 2026-01-14 6d62539
Add UI testing how-tos and explanation (#1066)
* Move docs images into docs/assets/images

* Add page on debugging UI test failures

* Add contributor documentation about ui testing and dependency security

* Check links job: Ignore links to medium blog posts

The same problem as `blog.jupyter.org` links. They seem to reject
requests that don't come from browsers.

* Lint :bell:

* Fixup roles with correct MyST syntax :bell:

* Fix incorrect link path

* Fixup broken crossrefs

* Eliminate an unnecessary heading level

* Remove title overrides from crossrefs

* Remove redundant titles

I was h
[…]
