### 2026-01-14 9df5a57
Update CartoVector layer type property to VectorTileLayer.
