### 2026-01-15 7024f63
Inline the function in useStacIndex
