### 2026-01-15 6f49f09
Update layerBrowserDialog component, the ILayerGalleryEntry interface and the createDefaultLayerRegistry function to have the proper logics to support both raster and vector tile layers.
