+++ packages/schema/src/interfaces.ts
@@ -29,8 +29,8 @@ import {
   IJGISStoryMap,
   SourceType,
 } from './_interface/project/jgis';
-import { IRasterSource } from './_interface/project/sources/rasterSource';
-import { Modes } from './types';
+
+import { IRasterSource, IVectorTileSource, Modes } from './types';
 export { IGeoJSONSource } from './_interface/project/sources/geoJsonSource';
 
 export interface IJGISStoryMaps {
@@ -351,19 +351,45 @@ export interface IJGISExternalCommandRegistry {
 }
 
 /**
- * Defines the structure for entries in a raster layer gallery.
- * Each entry consists of a name, a thumbnail URL, and source information.
- * The source information is expected to conform to the IRasterSource interface.
+ * Defines the base structure for entries in the layer gallery.
+ * Each entry consists of a name, a thumbnail URL, and a layer type information.
+ *
+ * @interface IBaseGalleryEntry
+ */
+interface IBaseGalleryEntry {
+  name: string;
+  thumbnail: string;
+  layerType: 'RasterLayer' | 'VectorTileLayer';
+}
+/**
+ * Defines the structure for entries of raster layers in the gallery.
+ * It extends IBaseGalleryEntry by providing a RasterSource type and a source expected to conform to the RasterSource interface.
+ *
+ * @interface IRasterGalleryEntry
+ */
+
+export interface IRasterGalleryEntry extends IBaseGalleryEntry {
+  sourceType: 'RasterSource';
+  source: IRasterSource;
+}
+
+/**
+ * Defines the structure for entries of vector tile layers in the gallery.
+ * It extends IBaseGalleryEntry by providing a VectorTileSource type and a source expected to conform to the VectorTileSource interface.
+ *
+ * @interface IRasterGalleryEntry
+ */
+export interface IVectorTileGalleryEntry extends IBaseGalleryEntry {
+  sourceType: 'VectorTileSource';
+  source: IVectorTileSource;
+}
+
+/**
+ * Defines the structure for entries in a layer gallery.
  *
  * @interface ILayerGalleryEntry
  */
-export interface ILayerGalleryEntry {
-  name: string;
-  thumbnail: string;
-  source: IRasterSource;
-  sourceType: any;
-  layerType: any;
-}
+export type ILayerGalleryEntry = IRasterGalleryEntry | IVectorTileGalleryEntry;
 
 export interface IJGISLayerBrowserRegistry {
   getRegistryLayers(): ILayerGalleryEntry[];
