### 2026-01-15 3d5f93c
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
