+++ packages/base/layer_gallery_generator.py
@@ -16,7 +16,7 @@ def fetch_tile(url_template, x, y, z, s="a"):
     Fetch a tile from the given URL template.
     """
     url = url_template.format(x=x, y=y, z=z, s=s)
-    #print(f"   Fetch {url}")
+    # print(f"   Fetch {url}")
     response = requests.get(
         url, headers={"Content-Type": "application/json", "User-Agent": "JupyterGIS"}
     )
@@ -83,59 +83,22 @@ middle_europe = {"lat": 48.63290858589535, "lng": -350.068359375, "zoom": 4}
 france = {"lat": 47.040182144806664, "lng": 1.2963867187500002, "zoom": 5}
 
 providers_types = {
-    "OpenStreetMap":{
-        "layerType": "RasterLayer",
-        "sourceType": "RasterSource"
-        },
-       
-        "NASAGIBS": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource"
-        },
-        "USGS": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource"
-        },
-        "WaymarkedTrails": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource"
-        },
-        "Gaode": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource"
-        },
-        "Strava": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource"
-        },
-        "OPNVKarte": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource"
-        },
-        "OpenTopoMap": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource"
-        },
-        "OpenRailwayMap": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource"
-        },
-        "Esri": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource"
-    },
-        "MacroStrat": {
-            "CartoRaster":{
-                "layerType": "RasterLayer",
-                "sourceType": "RasterSource"
-        },
-            "CartoRaster":{
-                "layerType": "RasterLayer",
-                "sourceType": "RasterSource"
-        },
-            "CartoVector":{
-                "layerType": "VectorTileLayer",
-                "sourceType": "VectorTileSource"
+    "OpenStreetMap": {"layerType": "RasterLayer", "sourceType": "RasterSource"},
+    "NASAGIBS": {"layerType": "RasterLayer", "sourceType": "RasterSource"},
+    "USGS": {"layerType": "RasterLayer", "sourceType": "RasterSource"},
+    "WaymarkedTrails": {"layerType": "RasterLayer", "sourceType": "RasterSource"},
+    "Gaode": {"layerType": "RasterLayer", "sourceType": "RasterSource"},
+    "Strava": {"layerType": "RasterLayer", "sourceType": "RasterSource"},
+    "OPNVKarte": {"layerType": "RasterLayer", "sourceType": "RasterSource"},
+    "OpenTopoMap": {"layerType": "RasterLayer", "sourceType": "RasterSource"},
+    "OpenRailwayMap": {"layerType": "RasterLayer", "sourceType": "RasterSource"},
+    "Esri": {"layerType": "RasterLayer", "sourceType": "RasterSource"},
+    "MacroStrat": {
+        "CartoRaster": {"layerType": "RasterLayer", "sourceType": "RasterSource"},
+        "CartoRaster": {"layerType": "RasterLayer", "sourceType": "RasterSource"},
+        "CartoVector": {
+            "layerType": "VectorTileLayer",
+            "sourceType": "VectorTileSource",
         },
     },
 }
[…]
