### 2026-01-16 d86b008
Update story toggle icon
