+++ packages/base/src/constants.ts
@@ -59,7 +59,9 @@ const iconObject = {
   [CommandIDs.temporalController]: { icon: clockIcon },
   [CommandIDs.addMarker]: { icon: markerIcon },
   [CommandIDs.addStorySegment]: { iconClass: 'fa fa-link' },
-  [CommandIDs.toggleStoryPresentationMode]: { iconClass: 'fa fa-dragon' },
+  [CommandIDs.toggleStoryPresentationMode]: {
+    iconClass: 'fa fa-book-open jgis-icon-adjust',
+  },
 };
 
 /**
