### 2026-01-16 64f2721
add setting for disable/enable zoom buttons
