### 2026-01-16 #9702 @crvernon
# Update regarding JOSS scope and paper format changes

## What this means for your submission

For full details, please see:
* Our updated scope definition: https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance
* An example paper in the new format: https://joss.readthedocs.io/en/latest/example_paper.html

We understand this may be frustrating, particularly if you've been waiting for some time. We apologize for that. These changes reflect the rapidly evolving landscape of software development and our commitment to ensuring JOSS publications recognize genuine scholarly contributions.
