### 2026-01-16 7ea0eeb
Change framing of setting to "enable", update default to `false`
We think of this setting as "opt-in", so it makes sense for the box to
initially be not checked. We want the default to be for the buttons to
be hidden, so we need to change the setting from "disable" to "enable"
to meet those conditions!

Co-authored-by: Nakul Verma <173621577+nakul-py@users.noreply.github.com>
