### 2026-01-18 a6b0ce0
Fix line 38 in useGenericStacSearch
Co-authored-by: Nakul Verma <nakulverma.py@gmail.com>
