### 2026-01-18 6db26ef
Fix line 115 in useGenericStacSearch
Co-authored-by: Nakul Verma <nakulverma.py@gmail.com>
