### 2026-01-18 36d5aab
fix(qgis_loader): replace auto-detection with opt-in JGIS_QGIS_PATH env var
