### 2026-01-18 89f00ca
Change to inline function in useStacIndex
