+++ packages/base/src/stacBrowser/hooks/useGenericStacSearch.ts
@@ -2,7 +2,6 @@ import { IJupyterGISModel, IJGISLayer } from '@jupytergis/schema';
 import { UUID } from '@lumino/coreutils';
 import { useEffect, useState } from 'react';
 
-
 import { fetchWithProxies } from '@/src/tools';
 import { IStacItem } from '../types/types';
 
