### 2026-01-19 6c771a1
Adding manual buttons for zoom-in or zoom-out (#1055)
* adding manual buttons for zoom-in or zoom-out

* add setting for disable/enable zoom buttons

* moving buttons to bottom right

* Change framing of setting to "enable", update default to `false`

We think of this setting as "opt-in", so it makes sense for the box to
initially be not checked. We want the default to be for the buttons to
be hidden, so we need to change the setting from "disable" to "enable"
to meet those conditions!

Co-authored-by: Nakul Verma <173621577+nakul-py@users.noreply.github.com>

* Adding icon for Jupytergis sett
[…]
