### 2026-01-23 c9a64bb
Disable toggle presentation button when no segments & change icon
