+++ packages/base/src/constants.ts
@@ -60,7 +60,7 @@ const iconObject = {
   [CommandIDs.addMarker]: { icon: markerIcon },
   [CommandIDs.addStorySegment]: { iconClass: 'fa fa-link' },
   [CommandIDs.toggleStoryPresentationMode]: {
-    iconClass: 'fa fa-book-open jgis-icon-adjust',
+    iconClass: 'fa fa-book jgis-icon-adjust',
   },
 };
 
