### 2026-01-23 8130194
Pin sphinx to 8.x to fix docs build
Failed build is using Sphinx 9.1.0: https://app.readthedocs.org/projects/jupytergis/builds/31116291/
Last passing build used Sphinx 8.2.3: https://app.readthedocs.org/projects/jupytergis/builds/31079760/

I don't know what caused the upgrade in the last 2 days; Sphinx 9 has
been out for over a month.
