+++ docs/environment-docs.yml
@@ -9,7 +9,7 @@ dependencies:
   - jupyterlite-core
   - jupyterlite-xeus
   - jupyterlite-sphinx
-  - sphinx
+  - sphinx ~=8.2.3
   - sphinx-tabs
   - pydata-sphinx-theme
   - sphinx-autodoc-typehints
