### 2026-01-23 5e14ad7
[WIP] Add stacindex.org catalog browsing interface (#1062)
Co-authored-by: Jessenia Chan <96598953+JesseniaChan@users.noreply.github.com>
Co-authored-by: Nakul Verma <nakulverma.py@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
