### 2026-01-26 ad04eac
Get rid of gifs
