### 2026-01-26 f66fabf
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
