### 2026-01-27 967eb86
Fix file loading
Taking the given file format into account
