### 2026-01-27 21aad1e
Add Story mode presentation toggle button to toolbar (#1056)
* Add Story mode presentation toggle button to toolbar

* Lint

* Update story toggle icon

* Update Playwright Snapshots

* Disable toggle presentation button when no segments & change icon

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
