### 2026-01-27 245be2e
Add a Story Map presentation mode for Specta (#1061)
* Add Story mode presentation toggle button to toolbar

* Lint

* Start specta UI changes

* Move OpenLayer controls to bottom right

* Story Viewer CSS

* Clean up a bit

* Fix stories loading

* Put title above image

* Hide top bars

* Disable commands in specta

* Only enable specta presentation mode when stories are present

* Remove top gap from panels

* Allow setting presentation bg and text color in editor panel

* Use RJSF types

* Get presentation colors from file

* CSS tweaks

* Fix syncStoryData

* Scrollytelling'

* CSS tweaki
[…]
