### 2026-01-27 cf9be92
Pin jupytext in docs build
