### 2026-01-28 1776d4e
Publish 0.12.0
SHA256 hashes:

jupytergis-0.12.0-py3-none-any.whl: 30a4b0675d869f6c7b822c839282d5c0a1119f90479873608786d5748f8c95a6

jupytergis-0.12.0.tar.gz: 6d73b3cb358695d0b6c4a0e79ad5f8cdfb69ad750a65cefd4e4f072486fa149a

jupytergis-base-0.12.0.tgz: c904f8cd25e1d5ca2c4284c75e6178e352d99d3d22bfc44c88210fdb1ef6b802

jupytergis-jupytergis-core-0.12.0.tgz: e712b43215d264e80c13d0b90579cd790037187c1c775f8218482fb47eec210d

jupytergis-jupytergis-lab-0.12.0.tgz: 561e160f5f426f6895b0aa735c40d12c53e961084153c1fb25af00aba79eedde

jupytergis-jupytergis-qgis-0.12.0.tgz: f25d8906ea1ca804c3e314488e2ad39db
[…]
