### 2026-01-28 169f732
Get back intellisense for commands
