+++ packages/base/src/constants.ts
@@ -1,4 +1,4 @@
-import { IDict, ProcessingCommandIDs } from '@jupytergis/schema';
+import { ProcessingCommandIDs } from '@jupytergis/schema';
 import { LabIcon, redoIcon, undoIcon } from '@jupyterlab/ui-components';
 
 import * as BaseCommandIDs from './commands/BaseCommandIDs';
@@ -16,8 +16,7 @@ import {
 /**
  * The command IDs.
  */
-
-export const CommandIDs: IDict = {
+export const CommandIDs: typeof BaseCommandIDs & typeof ProcessingCommandIDs = {
   ...BaseCommandIDs,
   ...ProcessingCommandIDs,
 };
