### 2026-01-28 25c0737
Extend the gallery of layer to support more layer types and sources.
