+++ packages/base/layer_gallery_generator.py
@@ -1,0 +1,618 @@
+from datetime import date, timedelta
+import json
+from io import BytesIO
+import os
+import subprocess
+
+import requests
+from PIL import Image
+import mercantile
+from xyzservices import providers, TileProvider
+import string
+from requests.exceptions import RequestException
+
+THUMBNAILS_LOCATION = "layer_gallery"
+
+def placeholder_tile(size):
+    return Image.new("RGB", size, (220, 220, 220))
+
+def extract_placeholders(url_template):
+    """
+    Extract the placeholders from an url_template
+    """
+    formatter = string.Formatter()
+    return {
+        field_name
+        for _, field_name, _, _ in formatter.parse(url_template)
+        if field_name
+    }
+
+def build_url_parameters(tile_provider):
+    """
+    Build the url parameters that are needed in fetch_tile, create_thumbnail and download_thumbnail
+    """
+    url_template = tile_provider["url"]
+    placeholders = extract_placeholders(url_template)
+
+    # Placeholders handled explicitly by fetch_tile
+    reserved = {"x", "y", "z", "s"}
+
+    kwargs = {}
+
+    for name in placeholders - reserved:
+        if name in tile_provider and tile_provider[name] is not None:
+            kwargs[name] = tile_provider[name]
+            if (name == "time" and tile_provider['time']==''):
+                kwargs['time']= yesterday
+        else:
+            raise KeyError(
+                f"Placeholder '{name}' not found in TileProvider '{tile_provider.get('name')}'"
+            )
+    return kwargs
+
+def fetch_tile(url_template, x, y, z, s="a", **kwargs):
+    """
+    Fetch a tile from the given URL template.
+    """
+    try:
+        url = url_template.format(x=x, y=y, z=z, s=s, **kwargs)
+        response = requests.get(
+            url,
+            headers={"User-Agent": "JupyterGIS"},
+            timeout=10,
+        )
+        response.raise_for_status()
+        return Image.open(BytesIO(response.content))
+    except RequestException as e:
+        print(f"⚠️ Tile fetch failed: {e}")
+        return None
+
+def latlng_to_tile(lat, lng, zoom):
+    """
+    Convert latitude/longitude to tile coordinates.
+    """
+    tile = mercantile.tile(lng, lat, zoom, True)
+    return tile.x, tile.y
+    
+
+
+def create_thumbnail(
+    url_template,
+    lat,
+    lng,
+    zoom,
+    thumbnail_path,
+    tile_size=256,
+    thumbnail_size=(512, 512),
+    **url_parameters,
+):
+    """
+    Create a thumbnail for the specified location and zoom level.
+    """
+    # Skip if thumbnail already exists
+    x, y = latlng_to_tile(lat, lng, zoom)
+
+    # Fetch the tiles (2x2 grid for the thumbnail)
+    tiles = []
+    for dy in range(2):
+        row = []
+        for dx in range(2):
+            tile_x, tile_y = x + dx, y + dy
+            tile = fetch_tile(url_template, tile_x, tile_y, zoom, **url_parameters)
+            if tile is None:
+                tile = placeholder_tile((tile_size, tile_size))
+            row.append(tile)
+        tiles.append(row)
+
+    # Create a blank image for the thumbnail
+    thumbnail = Image.new("RGB", (2 * tile_size, 2 * tile_size))
+
+    # Paste the tiles into the thumbnail image
+    for dy, row in enumerate(tiles):
+        for dx, tile in enumerate(row):
+            thumbnail.paste(tile, (dx * tile_size, dy * tile_size))
+
[…]
