### 2026-01-28 6e8711a
Extend the gallery of layer to support more layer types and sources.
