### 2026-01-28 d1cd417
Move provider_config to a json file.
