+++ packages/base/layer_gallery_generator.py
@@ -11,6 +11,10 @@ from xyzservices import providers, TileProvider
 import string
 from requests.exceptions import RequestException
 
+with open("layer_gallery/layer_config.json", "r", encoding="utf-8") as f:
+    provider_config = json.load(f)
+
+
 THUMBNAILS_LOCATION = "layer_gallery"
 
 def placeholder_tile(size):
@@ -122,323 +126,13 @@ yesterday = (date.today() - timedelta(days=1)).strftime("%Y-%m-%d")
 # San Francisco
 san_francisco = {"lat": 37.7749, "lng": -122.4194, "zoom": 5}
 
-middle_europe = {"lat": 48.63290858589535, "lng": -350.068359375, "zoom": 4}
+##middle_europe = {"lat": 48.63290858589535, "lng": -350.068359375, "zoom": 4}
 
 # Default
-france = {"lat": 47.040182144806664, "lng": 1.2963867187500002, "zoom": 5}
+#france = {"lat": 47.040182144806664, "lng": 1.2963867187500002, "zoom": 5}
 
-provider_config = {
-    "OpenStreetMap": {
-        "Mapnik": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource",
-            "thumbnail":{
-                "Special Rules": {},
-                "Default": france
-            },
-        },
-        "BZH": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource",
-            "thumbnail":{
-                "Special Rules": {"lat": 47.76702233051035, "lng": -3.4675598144531254, "zoom": 8},
-                "Default": france
-            },
-        },
-        "CAT": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource",
-            "thumbnail":{
-                "Special Rules": {},
-                "Default": france
-            },
-        },
-        "CH": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource",
-            "thumbnail":{
-                "Special Rules": {"lat": 46.8182, "lng": 8.2275, "zoom": 8},
-                "Default": france
-            },
-        },
-        "DE": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource",
-            "thumbnail":{
-                "Special Rules": {"lat": 51.1657, "lng": 10.4515, "zoom": 8},
-                "Default": france
-            }
-        },
-        "France": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource",
-            "thumbnail":{
-                "Special Rules": {},
-                "Default": france
-            }
-        },
-        "HOT": {
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource",
-            "thumbnail":{
-                "Special Rules": {},
-                "Default": france
-            },
-        }
-    },               
-    "NASAGIBS": {
-        "ModisTerraTrueColorCR" :{
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource",
-            "thumbnail": {
-                "Special Rules": {},
-                "Default": france,
-            }
-        },
-        "ModisTerraBands367CR" :{
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource",
-            "thumbnail": {
-                "Special Rules": {},
-                "Default": france,
-            }
-        },
-        "ViirsEarthAtNight2012" :{
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource",
-            "thumbnail": {
-                "Special Rules": {},
-                "Default": france,
-            }
-        },
-        "ModisTerraLSTDay" :{
-            "layerType": "RasterLayer",
-            "sourceType": "RasterSource",
-            "thumbnail": {
-                "Special Rules": {},
-                "Default": france,
-            }
-        },
-        "ModisTerraSnowCover" :{
-            "layerType": "RasterLayer",
[…]
