### 2026-01-29 d86b658
Move CSS var extraction and remove unused dependency
