### 2026-01-29 fe991c2
Set default minZoom to 0 when no value is provided by xyzservices.
