### 2026-01-30 daf759d
Take review comments into account. Add a vector layer with geoJSONSource in the layer_gallery.json file.
