### 2026-01-30 77bd094
Add the source type in the layer card used in the gallery.
