### 2026-01-30 2ceae28
Save current segment index in model
