### 2026-01-30 #9702 @ianhi
> [Ian Hunt-Isaak](https://reviewers.joss.theoj.org/reviewers/2886) would be a great fit to review this submission if he is available. He is actively involved in the Jupyter stack and the xarray ecosystem, both of which a closely with the focus of this work.

dependent on the result of the above message - I'd just like to note that I would be happy serve as a reviewer when that stage is reached
