### 2026-02-02 b31da74
Story map fixes (#1094)
* Switch to new segment when adding

* Use theme colors by default

* Notify command when segments populate

* Get back intellisense for commands

* Lint

* Typo

* Move CSS var extraction and remove unused dependency

* Lint?

* <_<

* >_>

* Fix overflow scroll on object properties panel

* Switch to story panel when entering presentation mode

* Move the shadow

* Update Playwright Snapshots

* Update Playwright Snapshots

* Fix Specta css

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-
[…]
