### 2026-02-02 7880b37
Publish 0.12.1
SHA256 hashes:

jupytergis-0.12.1-py3-none-any.whl: 2dccacd9e574f51ca41cb559750da1a870d19d935cf0107bcc3e7ef44fce1bf3

jupytergis-0.12.1.tar.gz: 8262fcb3fced9c19b0b76e978a0946fa3c5b24d4b9e7b7364dee70e0d518266f

jupytergis-base-0.12.1.tgz: da5118db02049115c3fbc56e104a41fd720837ec5cee697caa6a42073412f494

jupytergis-jupytergis-core-0.12.1.tgz: 73a56ef361c8c4e5d09d9d78a1d4621406ac59cbad391f03870a1b1198271551

jupytergis-jupytergis-lab-0.12.1.tgz: a2feeccac894381aa4c507b6f6230f0c3b7960145195b0fab9d459e06a740ec6

jupytergis-jupytergis-qgis-0.12.1.tgz: 9f027df8805f9f3894f92cb43483df68f
[…]
