### 2026-02-03 454c5e6
Select top layer by default (#1100)
* Select top layer by default

* Select layer when creating it

* Update Playwright Snapshots

* Update Playwright Snapshots

* Update test

* And anotha one

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
