### 2026-02-03 f020e1f
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
