+++ packages/base/layer_gallery_generator.py
@@ -16,6 +16,7 @@ with open("layer_gallery/thumbnail_config.json", "r", encoding="utf-8") as f:
 
 THUMBNAILS_LOCATION = "layer_gallery"
 
+
 def snake_to_camel(s):
     """
     Convect snake case strings into camel case ones
@@ -23,25 +24,25 @@ def snake_to_camel(s):
     parts = s.split("_")
     return parts[0] + "".join(word.capitalize() for word in parts[1:])
 
+
 def dict_keys_to_camel(obj):
     """
     Convect keys of a dict from snake case to camel case
     """
     if isinstance(obj, dict):
-        return {
-            snake_to_camel(k): dict_keys_to_camel(v)
-            for k, v in obj.items()
-        }
+        return {snake_to_camel(k): dict_keys_to_camel(v) for k, v in obj.items()}
     elif isinstance(obj, list):
         return [dict_keys_to_camel(item) for item in obj]
     else:
         return obj
+
 
 def placeholder_tile(size):
     """
     Provide a white empty placeholder image
     """
     return Image.new("RGB", size, (220, 220, 220))
+
 
 def extract_placeholders(url_template):
     """
@@ -53,6 +54,7 @@ def extract_placeholders(url_template):
         for _, field_name, _, _ in formatter.parse(url_template)
         if field_name
     }
+
 
 def build_url_parameters(tile_provider):
     """
@@ -69,13 +71,14 @@ def build_url_parameters(tile_provider):
     for name in placeholders - reserved:
         if name in tile_provider and tile_provider[name] is not None:
             kwargs[name] = tile_provider[name]
-            if (name == "time" and tile_provider['time']==''):
-                kwargs['time']= yesterday
+            if name == "time" and tile_provider["time"] == "":
+                kwargs["time"] = yesterday
         else:
             raise KeyError(
                 f"Placeholder '{name}' not found in TileProvider '{tile_provider.get('name')}'"
             )
     return kwargs
+
 
 def fetch_tile(url_template, x, y, z, s="a", **kwargs):
     """
@@ -94,13 +97,15 @@ def fetch_tile(url_template, x, y, z, s="a", **kwargs):
         print(f"⚠️ Tile fetch failed: {e}")
         return None
 
+
 def latlng_to_tile(lat, lng, zoom):
     """
     Convert latitude/longitude to tile coordinates.
     """
     tile = mercantile.tile(lng, lat, zoom, True)
     return tile.x, tile.y
-    
+
+
 def create_thumbnail(
     url_template,
     lat,
@@ -145,12 +150,12 @@ def create_thumbnail(
 yesterday = (date.today() - timedelta(days=1)).strftime("%Y-%m-%d")
 
 # San Francisco
-#san_francisco = {"lat": 37.7749, "lng": -122.4194, "zoom": 5}
+# san_francisco = {"lat": 37.7749, "lng": -122.4194, "zoom": 5}
 
 ##middle_europe = {"lat": 48.63290858589535, "lng": -350.068359375, "zoom": 4}
 
 # Default
-#france = {"lat": 47.040182144806664, "lng": 1.2963867187500002, "zoom": 5}
+# france = {"lat": 47.040182144806664, "lng": 1.2963867187500002, "zoom": 5}
 
 
 def download_thumbnail(url_template, name, position, tile_size, **url_parameters):
@@ -169,10 +174,11 @@ def download_thumbnail(url_template, name, position, tile_size, **url_parameters
     thumbnail.save(file_path)
     return file_path
 
+
 # Create thumbnail dir if needed
 if not os.path.exists(THUMBNAILS_LOCATION):
[…]
