### 2026-02-03 9685816
Update layer gallery to support more layer types and sources (#1071)
* Extend the gallery of layer to support more layer types and sources.

* Move provider_config to a json file.

* Set default minZoom to 0 when no value is provided by xyzservices.

* Take review comments into account. Add a vector layer with geoJSONSource in the layer_gallery.json file.

* Add the source type in the layer card used in the gallery.

* Restore WaymarkedTrails layers.

* Move urlParameters into sourceParameters to properly fit the schema.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more informa
[…]
