### 2026-02-04 df24fff
Rename slide -> segment
