+++ packages/schema/src/interfaces.ts
@@ -284,8 +284,8 @@ export interface IJupyterGISModel extends DocumentRegistry.IModel {
     storyId: string;
     story: IJGISStoryMap | undefined;
   };
-  getCurrentSlideIndex(): number;
-  setCurrentSlideIndex(index: number): void;
+  getCurrentSegmentIndex(): number;
+  setCurrentSegmentIndex(index: number): void;
   addStorySegment(): { storySegmentId: string; storyId: string } | null;
   segmentAdded: ISignal<
     IJupyterGISModel,
