### 2026-02-04 4ab38b6
Add mobile-friendly story map view for specta (#1096)
* Update radix

* Start setup for fancy mobile view

* Refactor to use slot for NavBar

* Add blank segment

* Get Navbar looking right on specta mobile

* Use story colors in mobile specta

* Hide status bar in specta

* Move the button

* Save current segment index in model

* Only move map when segment actually changes

* Not bad

* Fin

* lol yo really?

* Clean up

* Rename slide -> segment
