### 2026-02-04 a0b8553
Publish 0.12.2
SHA256 hashes:

jupytergis-0.12.2-py3-none-any.whl: 922305558cb2c4f8e02fa517dcf20c3e096bd3be810fa2e8209c53a5be7a5eae

jupytergis-0.12.2.tar.gz: cbc353769d606cb9686520e22b9850d798ff59d3b308c9b3f212b4fd1dc43936

jupytergis-base-0.12.2.tgz: a40bc847e67fdfdaf4c80b8a168a0b1ec490f40c7fe63f66cc688bfdd10710d9

jupytergis-jupytergis-core-0.12.2.tgz: 6b177ff528ea464e79497e867d02908d561caf511df06d5d300a34b3592a15a6

jupytergis-jupytergis-lab-0.12.2.tgz: 8d64bf98eac54298b5a979b7c8b273ff2bdc3ee40823a08bbd211361e6afd162

jupytergis-jupytergis-qgis-0.12.2.tgz: 3c520de7f1c0e756a128bbaa462373b17
[…]
