+++ python/jupytergis/jupytergis/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.12.1"
+__version__ = "0.12.2"
 
 from jupytergis_lab import GISDocument, explore  # noqa
