### 2026-02-05 74574f1
Update notebook.link lock with mambajs (#1103)
* Update lock with mambajs

* Add docs on how to do it

* Up
