+++ packages/base/src/shared/hooks/useLatest.ts
@@ -1,0 +1,11 @@
+import { useEffect, useRef } from 'react';
+
+export function useLatest<T>(value: T): React.MutableRefObject<T> {
+  const ref = useRef(value);
+
+  useEffect(() => {
+    ref.current = value;
+  }, [value]);
+
+  return ref;
+}
