+++ packages/base/src/dialogs/symbology/symbologyUtils.ts
@@ -1,18 +1,160 @@
-import { IJGISLayer } from '@jupytergis/schema';
+import {
+  IJGISLayer,
+  IJupyterGISModel,
+  IVectorLayer,
+  IWebGlLayer,
+} from '@jupytergis/schema';
 import colormap from 'colormap';
 
 import { ColorRampName } from './colorRampUtils';
 import { IStopRow } from './symbologyDialog';
 
 const COLOR_EXPR_STOPS_START = 3;
+
+/** Payload when saving symbology; shape matches vector or WebGl layer params. */
+export interface ISymbologyPayload {
+  symbologyState:
+    | IVectorLayer['symbologyState']
+    | IWebGlLayer['symbologyState'];
+  color?: IVectorLayer['color'] | IWebGlLayer['color'];
+}
+
+export interface ISaveSymbologyOptions {
+  model: IJupyterGISModel;
+  layerId: string;
+  isStorySegmentOverride?: boolean;
+  segmentId?: string;
+  payload: ISymbologyPayload;
+  mutateLayerBeforeSave?: (layer: any) => void;
+}
+
+export type VectorSymbologyParams = Pick<
+  IVectorLayer,
+  'symbologyState' | 'color'
+>;
+
+export type WebGlSymbologyParams = Pick<
+  IWebGlLayer,
+  'symbologyState' | 'color'
+>;
+
+/** Params-shaped object used for reading symbology (layer.parameters or segment override). */
+export type IEffectiveSymbologyParams =
+  | VectorSymbologyParams
+  | WebGlSymbologyParams;
+
+/**
+ * Resolve the effective symbology params for this dialog: either the layer's
+ * parameters or the matching segment override when editing a story-segment override.
+ */
+export function getEffectiveSymbologyParams(
+  model: IJupyterGISModel,
+  layerId: string,
+  layer: IJGISLayer | null | undefined,
+  isStorySegmentOverride?: boolean,
+  segmentId?: string,
+): IEffectiveSymbologyParams | null {
+  if (!layer?.parameters) {
+    return null;
+  }
+  if (!isStorySegmentOverride) {
+    return layer.parameters as IEffectiveSymbologyParams;
+  }
+  if (!segmentId) {
+    return null;
+  }
+  const segment = model.getLayer(segmentId);
+  const override = segment?.parameters?.symbologyOverride?.find(
+    (override: { targetLayer?: string }) => override.targetLayer === layerId,
+  );
+
+  if (!override.symbologyState) {
+    override.symbologyState = {};
+  }
+  return (override as IEffectiveSymbologyParams) ?? null;
+}
+
+export function saveSymbology(options: ISaveSymbologyOptions): void {
+  const {
+    model,
+    layerId,
+    isStorySegmentOverride,
+    segmentId,
+    payload,
+    mutateLayerBeforeSave,
+  } = options;
+
+  if (!isStorySegmentOverride) {
+    const layer = model.getLayer(layerId);
+    if (!layer?.parameters) {
+      return;
+    }
+
+    layer.parameters.symbologyState = payload.symbologyState;
+    if (payload.color !== undefined) {
+      layer.parameters.color = payload.color;
+    }
+
+    mutateLayerBeforeSave?.(layer);
+    model.sharedModel.updateLayer(layerId, layer);
+    return;
+  }
+
+  if (!segmentId) {
+    return;
+  }
+
+  const segment = model.getLayer(segmentId);
+  if (!segment?.parameters) {
+    return;
+  }
+
+  if (!segment.parameters.symbologyOverride) {
+    segment.parameters.symbologyOverride = [];
+  }
+
+  // Find the override for the target layer (from the selected layer in the dialog)
+  const targetLayerId = model.selected
+    ? Object.keys(model.selected).find(
+        id =>
[…]
