### 2026-02-05 9f7007e
Add symbologyState in the VectorTileLayer schema. Update layer_gallery_generator.py accordingly.
