+++ packages/base/layer_gallery_generator.py
@@ -219,7 +219,11 @@ for provider_key, provider_value in provider_config.items():
         file_path = download_thumbnail(
             url_template, provider_key, position, tile_size, **url_parameters
         )
-
+        if provider_value["layerType"] == "VectorTileLayer":
+            layerParameters = {"opacity": 1, "symbologyState": {}}
+        else:
+            layerParameters = {"opacity": 1
+            }
         provider_gallery[provider_key] = {
             "thumbnailPath": file_path,
             "name": provider_key,
@@ -232,9 +236,10 @@ for provider_key, provider_value in provider_config.items():
                 "minZoom": xyzprovider.get("min_zoom") or 0,
                 "urlParameters": dict_keys_to_camel(url_template),
             },
-            "layerParameters": {"opacity": 1},
+            "layerParameters": layerParameters,
         }
 
+    
     elif config_is_flat and not xyz_is_flat:
         providers_maps = {}
 
@@ -253,7 +258,12 @@ for provider_key, provider_value in provider_config.items():
             file_path = download_thumbnail(
                 url_template, name, position, tile_size, **url_parameters
             )
-
+            
+            if provider_value["layerType"] == "VectorTileLayer":
+                layerParameters = {"opacity": 1, "symbologyState": {}}
+            else:
+                 layerParameters = {"opacity": 1}
+                 
             providers_maps[map_name] = {
                 "thumbnailPath": file_path,
                 "name": provider_key + "." + map_name,
@@ -266,10 +276,10 @@ for provider_key, provider_value in provider_config.items():
                     "minZoom": tile_provider.get("min_zoom") or 0,
                     "urlParameters": dict_keys_to_camel(url_parameters),
                 },
-                "layerParameters": {"opacity": 1},
+                "layerParameters": layerParameters,
                 "description": tile_provider.get("attribution"),
             }
-
+            
         provider_gallery[provider_key] = providers_maps
 
     elif not config_is_flat and not xyz_is_flat:
@@ -290,6 +300,11 @@ for provider_key, provider_value in provider_config.items():
             file_path = download_thumbnail(
                 url_template, name, position, tile_size, **url_parameters
             )
+            
+            if map_config["layerType"] == "VectorTileLayer":
+                layerParameters = {"opacity": 1, "symbologyState": {}}
+            else:
+                 layerParameters = {"opacity": 1}
 
             providers_maps[map_name] = {
                 "thumbnailPath": file_path,
@@ -303,9 +318,10 @@ for provider_key, provider_value in provider_config.items():
                     "minZoom": tile_provider.get("min_zoom") or 0,
                     "urlParameters": dict_keys_to_camel(url_parameters),
                 },
-                "layerParameters": {"opacity": 1},
+                "layerParameters": layerParameters,
                 "description": tile_provider.get("attribution"),
             }
+            
 
         provider_gallery[provider_key] = providers_maps
 
