### 2026-02-05 93b6bc0
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
