+++ packages/base/layer_gallery_generator.py
@@ -222,8 +222,7 @@ for provider_key, provider_value in provider_config.items():
         if provider_value["layerType"] == "VectorTileLayer":
             layerParameters = {"opacity": 1, "symbologyState": {}}
         else:
-            layerParameters = {"opacity": 1
-            }
+            layerParameters = {"opacity": 1}
         provider_gallery[provider_key] = {
             "thumbnailPath": file_path,
             "name": provider_key,
@@ -239,7 +238,6 @@ for provider_key, provider_value in provider_config.items():
             "layerParameters": layerParameters,
         }
 
-    
     elif config_is_flat and not xyz_is_flat:
         providers_maps = {}
 
@@ -258,12 +256,12 @@ for provider_key, provider_value in provider_config.items():
             file_path = download_thumbnail(
                 url_template, name, position, tile_size, **url_parameters
             )
-            
+
             if provider_value["layerType"] == "VectorTileLayer":
                 layerParameters = {"opacity": 1, "symbologyState": {}}
             else:
-                 layerParameters = {"opacity": 1}
-                 
+                layerParameters = {"opacity": 1}
+
             providers_maps[map_name] = {
                 "thumbnailPath": file_path,
                 "name": provider_key + "." + map_name,
@@ -279,7 +277,7 @@ for provider_key, provider_value in provider_config.items():
                 "layerParameters": layerParameters,
                 "description": tile_provider.get("attribution"),
             }
-            
+
         provider_gallery[provider_key] = providers_maps
 
     elif not config_is_flat and not xyz_is_flat:
@@ -300,11 +298,11 @@ for provider_key, provider_value in provider_config.items():
             file_path = download_thumbnail(
                 url_template, name, position, tile_size, **url_parameters
             )
-            
+
             if map_config["layerType"] == "VectorTileLayer":
                 layerParameters = {"opacity": 1, "symbologyState": {}}
             else:
-                 layerParameters = {"opacity": 1}
+                layerParameters = {"opacity": 1}
 
             providers_maps[map_name] = {
                 "thumbnailPath": file_path,
@@ -321,7 +319,6 @@ for provider_key, provider_value in provider_config.items():
                 "layerParameters": layerParameters,
                 "description": tile_provider.get("attribution"),
             }
-            
 
         provider_gallery[provider_key] = providers_maps
 
