### 2026-02-06 dcacfd8
Add symbolologyState to the VectorTileLayer schema (#1106)
* Add symbologyState in the VectorTileLayer schema.
Update layer_gallery_generator.py accordingly.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
