### 2026-02-06 b148190
Rename symbologyOverride -> layerOverride
