### 2026-02-06 e1850c1
Add story segment symbology overrides (#1102)
* Squashed

* Fix rebase

* Lint

* Remove unused props

* Remove example

* Add story maps doc

* Add Short Screencast

* Remove unused css

* Allow selecting all layers

* Rename symbologyOverride -> layerOverride

* Use const in command
