### 2026-02-06 a524b2f
Publish 0.13.0
SHA256 hashes:

jupytergis-0.13.0-py3-none-any.whl: 6720dbe730951a3dde58bae5ff42db9b7db006b743c5b5685e46d366498f3489

jupytergis-0.13.0.tar.gz: 5b032330317fb17d714d090044b4b5196e754ace1c6165f73e7b635c7690b13e

jupytergis-base-0.13.0.tgz: bda87e34bb4214732f5a8e58e8230988c205623331aa248b41fe72854245eceb

jupytergis-jupytergis-core-0.13.0.tgz: ff60a2caac746e7bb690bb89d0182db6463159c72b6ab166a07d57619ee28fa6

jupytergis-jupytergis-lab-0.13.0.tgz: afe6b4236c43c937eb8e392d587687535715b38a7d8d28fa88250b10ab14cd3f

jupytergis-jupytergis-qgis-0.13.0.tgz: a01aee0e4f112e4f0acc17b6f60c8bceb
[…]
