+++ CITATION.cff
@@ -23,8 +23,8 @@ contact:
   - name: "The GeoJupyter Community Chat"
     website: "https://jupyter.zulipchat.com"
 
-version: "0.12.2"
-date-released: "2026-02-04"
+version: "0.13.0"
+date-released: "2026-02-06"
 
 # TODO: More comprehensive authors list. Initialized from the JOSS paper.
 authors:
