### 2026-02-09 e016600
Add empty string as path when creating gejson via python
