+++ packages/base/src/formbuilder/objectform/source/geojsonsource.ts
@@ -26,10 +26,10 @@ export class GeoJSONSourcePropertiesForm extends PathBasedSourcePropertiesForm {
     schema: IDict,
     uiSchema: IDict,
   ) {
-    if (data?.path !== '') {
-      this.removeFormEntry('data', data, schema, uiSchema);
-    }
+    this.removeFormEntry('data', data, schema, uiSchema);
+
     if (this.props.formContext === 'create') {
+
       schema.properties.path.description =
         'The local path to a GeoJSON file. (If no path/url is provided, an empty GeoJSON is created.)';
     }
