### 2026-02-09 f1734d8
Use defaults when creating layer in python
