+++ python/jupytergis_lab/jupytergis_lab/notebook/gis_document.py
@@ -285,8 +285,6 @@ class GISDocument(CommWidget):
                 # The front-end would have no way of finding the file reliably
                 with open(path, "r") as fobj:
                     parameters["data"] = json.load(fobj)
-                    # This is just to make rjsf happy
-                    parameters["path"] = ""
 
         if data is not None:
             parameters["data"] = data
@@ -931,10 +929,9 @@ class ObjectFactoryManager(metaclass=SingletonMeta):
         filters = data.get("filters", None)
         if object_type and object_type in self._factories:
             Model = self._factories[object_type]
-            args = {}
             params = data["parameters"]
-            for field in Model.__fields__:
-                args[field] = params.get(field, None)
+            # Only pass params that are present so Pydantic uses schema defaults for the rest
+            args = {k: params[k] for k in Model.model_fields if k in params}
             obj_params = Model(**args)
             return JGISLayer(
                 parent=parent,
@@ -954,10 +951,9 @@ class ObjectFactoryManager(metaclass=SingletonMeta):
         name: str = data.get("name", None)
         if object_type and object_type in self._factories:
             Model = self._factories[object_type]
-            args = {}
             params = data["parameters"]
-            for field in Model.__fields__:
-                args[field] = params.get(field, None)
+            # Only pass params that are present so Pydantic uses schema defaults for the rest
+            args = {k: params[k] for k in Model.model_fields if k in params}
             obj_params = Model(**args)
             return JGISSource(
                 parent=parent, name=name, type=object_type, parameters=obj_params
