### 2026-02-09 3770d27
Expand JupyterGIS JOSS submission to follow the updated guidelines
