+++ paper/paper.bib
@@ -8,6 +8,41 @@
   url = {https://qgis.org}
 }
 
+@software{CONDAFORGE2025,
+  author = {conda-forge community},
+  title = {The conda-forge Project: Community-based Software Distribution Built on the conda Package Format and Ecosystem},
+  year = 2015,
+  publisher = {Zenodo},
+  doi = {10.5281/zenodo.4774217},
+  url = {https://conda-forge.org},
+}
+
+@article{OPENEO2021,
+  title = {The openEO API–Harmonising the Use of Earth Observation Cloud Services Using Virtual Data Cube Functionalities},
+  volume = {13},
+  ISSN = {2072-4292},
+  url = {http://dx.doi.org/10.3390/rs13061125},
+  DOI = {10.3390/rs13061125},
+  number = {6},
+  journal = {Remote Sensing},
+  publisher = {MDPI AG},
+  author = {Schramm,  Matthias and Pebesma,  Edzer and Milenković,  Milutin and Foresta,  Luca and Dries,  Jeroen and Jacob,  Alexander and Wagner,  Wolfgang and Mohr,  Matthias and Neteler,  Markus and Kadunc,  Miha and Miksa,  Tomasz and Kempeneers,  Pieter and Verbesselt,  Jan and G\"{o}ßwein,  Bernhard and Navacchi,  Claudio and Lippens,  Stefaan and Reiche,  Johannes},
+  year = {2021},
+  month = mar,
+  pages = {1125}
+}
+
+@InProceedings{CRDT2011,
+  author="Shapiro, Marc and Pregui{\c{c}}a, Nuno and Baquero, Carlos and Zawirski, Marek",
+  editor="D{\'e}fago, Xavier and Petit, Franck and Villain, Vincent",
+  title="Conflict-Free Replicated Data Types",
+  booktitle="Stabilization, Safety, and Security of Distributed Systems",
+  year="2011",
+  publisher="Springer Berlin Heidelberg",
+  address="Berlin, Heidelberg",
+  pages="386--400",
+}
+
 @inproceedings{YJS2015,
   author = {Kevin Jahns and others},
   title = {Yjs: A Framework for Near Real-Time P2P Shared Editing on Arbitrary Data Types},
@@ -16,13 +51,55 @@
   doi = {10.1007/978-3-319-19890-3_55}
 }
 
-@inproceedings{JupyterNotebooks2018,
+@inproceedings{JUPY2018,
   author = {Kluyver, Thomas and others},
   title = {Jupyter Notebooks—a publishing format for reproducible computational workflows},
   booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
   year = {2018},
   publisher = {IOS Press},
   doi = {10.3233/978-1-61499-649-1-87}
+}
+
+@inproceedings{BINDER2018,
+  series = {SciPy},
+  title = {Binder 2.0 - Reproducible,  interactive,  sharable environments for science at scale},
+  ISSN = {2575-9752},
+  url = {http://dx.doi.org/10.25080/Majora-4af1f417-011},
+  DOI = {10.25080/majora-4af1f417-011},
+  booktitle = {Proceedings of the 17th Python in Science Conference},
+  publisher = {SciPy},
[…]
