### 2026-02-10 b40f36c
Missed some renames (#1108)
