### 2026-02-10 0e5cfb4
Remove old type stuff
