+++ python/jupytergis_lab/jupytergis_lab/notebook/gis_document.py
@@ -555,7 +555,6 @@ class GISDocument(CommWidget):
         self,
         path: str,
         name: str = "GeoParquetLayer",
-        type: "circle" | "fill" | "line" = "line",
         opacity: float = 1,
         logical_op: str | None = None,
         feature: str | None = None,
@@ -568,7 +567,6 @@ class GISDocument(CommWidget):
 
         :param path: The path to the GeoParquet file to embed into the jGIS file.
         :param name: The name that will be used for the object in the document.
-        :param type: The type of the vector layer to create.
         :param opacity: The opacity, between 0 and 1.
         :param logical_op: The logical combination to apply to filters. Must be "any" or "all"
         :param feature: The feature to be filtered on
@@ -591,7 +589,6 @@ class GISDocument(CommWidget):
             "visible": True,
             "parameters": {
                 "source": source_id,
-                "type": type,
                 "opacity": opacity,
                 "color": color_expr,
             },
