+++ python/jupytergis_qgis/jupytergis_qgis/tests/test_qgis.py
@@ -308,6 +308,7 @@ def test_qgis_saver():
                         "renderType": "Categorized",
                         "value": "min_label",
                     },
+                    "type": "line",
                 },
                 "type": "VectorLayer",
                 "visible": True,
