### 2026-02-10 6e162d9
Fix heatmap creation
