+++ docs/environment-docs.yml
@@ -4,6 +4,7 @@ channels:
   - nodefaults
 dependencies:
   - python
+  - setuptools
 
   # Build docs & JupyterLite
   - jupyterlite-core
