+++ docs/contributor_guide/development_setup.md
@@ -21,9 +21,7 @@ cd jupytergis
 ```
 
 ### Create a virtual environment
-
-```````{tabs}
-``````{group-tab} Micromamba (Recommended)
+``````{tab} Micromamba (Recommended)
 ```{note}
 Micromamba is a lightweight package manager compatible with conda environments.
 It is recommended for setting up the JupyterGIS development environment. If you don’t have it installed, please follow the official documentation: [Micromamba Installation Guide](https://mamba.readthedocs.io/en/latest/installation/micromamba-installation.html)
@@ -38,7 +36,7 @@ micromamba create --name jupytergis_dev -c conda-forge pip "nodejs<22" qgis
 micromamba activate jupytergis_dev
 ``````
 
-``````{group-tab} Plain python
+``````{tab} Plain python
 ```{note}
 You may need to install some non-Python dependencies (e.g. QGIS,
 Node.js) separately when using this method.
@@ -52,7 +50,6 @@ python -m venv .venv
 source .venv/bin/activate
 ```
 ``````
-```````
 
 ### Install dependencies and build
 
