+++ docs/user_guide/install.md
@@ -1,6 +1,6 @@
 # Installing JupyterGIS
 
-``````{tab} Mamba (recommended)
+````{tab} Mamba (recommended)
 
 :::{note}
 It is best to install JupyterGIS using `mamba`, `conda`, `micromamba`, or `pixi` since you'll be able to install `qgis` at the same time, allowing you to open `.qgz` files.
@@ -9,9 +9,9 @@ It is best to install JupyterGIS using `mamba`, `conda`, `micromamba`, or `pixi`
 ```bash
 mamba install -c conda-forge jupytergis qgis
 ```
-``````
+````
 
-``````{tab} pip
+````{tab} pip
 
 :::{warning}
 When installing with `pip`, QGIS compatibility functions won't work out of the box.
@@ -21,7 +21,7 @@ You're on your own!
 ```bash
 pip install jupytergis
 ```
-``````
+````
 
 Once JupyterGIS is installed, start JupyterLab:
 
@@ -33,27 +33,27 @@ jupyter lab
 
 Here are a few neat options to quickstart in an ephemeral environment:
 
-``````{tab} Docker
+````{tab} Docker
 
 [Docker](https://www.docker.com/)
 
 ```bash
 docker run -p 8888:8888 ghcr.io/geojupyter/jupytergis:latest
 ```
-``````
+````
 
-``````{tab} Pixi
+````{tab} Pixi
 [Pixi](https://github.com/prefix-dev/pixi)
 
 ```bash
 pixi exec --spec jupytergis --spec qgis jupyter lab
 ```
-``````
+````
 
-``````{tab} uv
+````{tab} uv
 [uv](https://github.com/astral-sh/uv)
 
 ```bash
 uv run --with jupytergis jupyter lab
 ```
-``````
+````
