### 2026-02-12 2e4c497
Publish 0.13.1
SHA256 hashes:

jupytergis-0.13.1-py3-none-any.whl: 6358f33a6f3354a6955dae7fd21e50cb989f1da416079154f782d38a7a95a172

jupytergis-0.13.1.tar.gz: 320322c37c7a1b35399c0e47d9cb5b023307c1e97d53bbe9377aef1e307be395

jupytergis-base-0.13.1.tgz: c54a6a7f52b5bc64a3a530caa127e6c19b710b9381be90f19ef4b80b2166701b

jupytergis-jupytergis-core-0.13.1.tgz: 9581fba4bc9fe0d941c72af0eacaa592dd0334933cfd92ab8b08737ef9ef29e6

jupytergis-jupytergis-lab-0.13.1.tgz: 39acd3a699329caaf572f3ac305941b2768e791c0e0a5ebb5580dc011f4811f0

jupytergis-jupytergis-qgis-0.13.1.tgz: 6249406be596bd65b5f0970c38a0d19df
[…]
