### 2026-02-14 7278fe9
stac-browser test fix
