### 2026-02-16 519813d
Fix openjournals/openjournals-draft-action action version
