### 2026-02-16 #9702 @SylvainCorlay
@crvernon I have updated the paper on the main branch of the JupyterGIS repository to follow the updated guidelines!

@ianhi thank you for volunteering.
