### 2026-02-17 83e9763
refactor remove layer/group command
