### 2026-02-18 5541db1
only remove/rename buttons for group
