+++ python/jupytergis_lab/src/index.ts
@@ -58,6 +58,8 @@ const plugin: JupyterFrontEndPlugin<void> = {
       );
     };
 
+    const LAYER = '.jp-gis-layerItem:not(.jp-gis-layerGroup)';
+
     createDefaultLayerRegistry(layerBrowserRegistry);
     const stateDbManager = GlobalStateDbManager.getInstance();
     stateDbManager.initialize(state);
@@ -87,32 +89,32 @@ const plugin: JupyterFrontEndPlugin<void> = {
     // LAYERS and LAYER GROUPS context menu
     app.contextMenu.addItem({
       command: CommandIDs.symbology,
-      selector: '.jp-gis-layerItem',
+      selector: LAYER,
       rank: 1,
     });
 
     // Separator
     app.contextMenu.addItem({
       type: 'separator',
-      selector: '.jp-gis-layerPanel',
+      selector: LAYER,
       rank: 1,
     });
 
     app.contextMenu.addItem({
       command: CommandIDs.removeSelected,
-      selector: '.jp-gis-layerItem',
+      selector: LAYER,
       rank: 2,
     });
 
     app.contextMenu.addItem({
       command: CommandIDs.renameSelected,
-      selector: '.jp-gis-layerItem',
+      selector: LAYER,
       rank: 2,
     });
 
     app.contextMenu.addItem({
       command: CommandIDs.zoomToLayer,
-      selector: '.jp-gis-layerItem',
+      selector: LAYER,
       rank: 2,
     });
 
@@ -128,7 +130,7 @@ const plugin: JupyterFrontEndPlugin<void> = {
     // Add the Download submenu to the context menu
     app.contextMenu.addItem({
       type: 'submenu',
-      selector: '.jp-gis-layerItem',
+      selector: LAYER,
       rank: 2,
       submenu: downloadSubmenu,
     });
@@ -148,7 +150,7 @@ const plugin: JupyterFrontEndPlugin<void> = {
 
     app.contextMenu.addItem({
       type: 'submenu',
-      selector: '.jp-gis-layerItem',
+      selector: LAYER,
       rank: 2,
       submenu: processingSubmenu,
     });
@@ -161,7 +163,7 @@ const plugin: JupyterFrontEndPlugin<void> = {
 
     app.contextMenu.addItem({
       type: 'submenu',
-      selector: '.jp-gis-layerItem',
+      selector: LAYER,
       rank: 2,
       submenu: moveLayerSubmenu,
     });
