### 2026-02-18 8acf370
Add Specta to env
