+++ lite/environment.yml
@@ -12,7 +12,6 @@ dependencies:
   - comm>=0.1.2,<0.2.0
   - pydantic>=2,<3
   - sidecar
-  - specta
   - pip:
       - yjs-widgets>=0.4,<0.5
       - my-jupyter-shared-drive>=0.2.2
