### 2026-02-18 0dcdde0
Focusing Layer/Group always to rename (#1123)
* Focusing layer everytime to rename

* renaming only on selected item

* refactor remove layer/group command

* only remove/rename buttons for group
