+++ packages/base/src/commands/index.ts
@@ -7,7 +7,6 @@ import {
   IJupyterGISModel,
   JgisCoordinates,
   LayerType,
-  SelectionType,
   SourceType,
 } from '@jupytergis/schema';
 import { JupyterFrontEnd } from '@jupyterlab/application';
@@ -534,41 +533,41 @@ export function addCommands(
   /**
    * LAYERS and LAYER GROUP actions.
    */
-  commands.addCommand(CommandIDs.renameLayer, {
-    label: trans.__('Rename Layer'),
+  commands.addCommand(CommandIDs.renameSelected, {
+    label: trans.__('Rename'),
+    isEnabled: () => {
+      const model = tracker.currentWidget?.model;
+      const selected = model?.localState?.selected?.value;
+      return !!selected && Object.keys(selected).length === 1;
+    },
     execute: async () => {
       const model = tracker.currentWidget?.model;
-      await Private.renameSelectedItem(model, 'layer');
+      const selected = model?.localState?.selected?.value;
+
+      if (!model || !selected) {
+        return;
+      }
+
+      await Private.renameSelectedItem(model);
     },
   });
 
-  commands.addCommand(CommandIDs.removeLayer, {
-    label: trans.__('Remove Layer'),
-    execute: () => {
+  commands.addCommand(CommandIDs.removeSelected, {
+    label: trans.__('Remove'),
+    isEnabled: () => {
       const model = tracker.currentWidget?.model;
-      Private.removeSelectedItems(model, 'layer', selection => {
-        model?.removeLayer(selection);
-      });
-
-      commands.notifyCommandChanged(CommandIDs.toggleStoryPresentationMode);
+      const selected = model?.localState?.selected?.value;
+      return !!selected && Object.keys(selected).length > 0;
     },
-  });
-
-  commands.addCommand(CommandIDs.renameGroup, {
-    label: trans.__('Rename Group'),
     execute: async () => {
       const model = tracker.currentWidget?.model;
-      await Private.renameSelectedItem(model, 'group');
-    },
-  });
+      const selected = model?.localState?.selected?.value;
 
-  commands.addCommand(CommandIDs.removeGroup, {
-    label: trans.__('Remove Group'),
-    execute: async () => {
-      const model = tracker.currentWidget?.model;
-      Private.removeSelectedItems(model, 'group', selection => {
-        model?.removeLayerGroup(selection);
-      });
+      if (!model || !selected) {
+        return;
+      }
+
+      await Private.removeSelectedItems(model);
     },
   });
 
@@ -659,7 +658,7 @@ export function addCommands(
     label: trans.__('Rename Source'),
     execute: async () => {
       const model = tracker.currentWidget?.model;
-      await Private.renameSelectedItem(model, 'source');
+      await Private.renameSelectedItem(model);
     },
   });
 
@@ -667,16 +666,7 @@ export function addCommands(
     label: trans.__('Remove Source'),
     execute: () => {
       const model = tracker.currentWidget?.model;
-      Private.removeSelectedItems(model, 'source', selection => {
-        if (!(model?.getLayersBySource(selection).length ?? true)) {
-          model?.sharedModel.removeSource(selection);
-        } else {
-          showErrorMessage(
-            'Remove source error',
-            'The source is used by a layer.',
-          );
-        }
-      });
+      Private.removeSelectedSources(model);
     },
   });
 
@@ -1230,52 +1220,71 @@ namespace Private {
     };
   }
 
-  export function removeSelectedItems(
-    model: IJupyterGISModel | undefined,
-    itemTypeToRemove: SelectionType,
-    removeFunction: (id: string) => void,
-  ) {
+  export function removeSelectedItems(model: IJupyterGISModel | undefined) {
     const selected = model?.localState?.selected?.value;
 
-    if (!selected) {
[…]
