### 2026-02-18 b067158
Publish 0.13.2
SHA256 hashes:

jupytergis-0.13.2-py3-none-any.whl: ea1a41d027a61b1013e6893897e615d875380447e1e723077a8a20291e7a9ac6

jupytergis-0.13.2.tar.gz: 0354485eb23da15088947680fad925958aa7cf22c78ed97f51c73268e33b372d

jupytergis-base-0.13.2.tgz: affbf7b4a0b95ad6d237fa0b2496afd0374bb5db3375013c46af22ebc30d42fd

jupytergis-jupytergis-core-0.13.2.tgz: 3eca3698a333b6e849b5ac3aeab9fc179d52d7e10707b861e3e07df8859eb934

jupytergis-jupytergis-lab-0.13.2.tgz: e5a9204f44a411741f0e18694da5e708155d12b129044915baba9c196783540b

jupytergis-jupytergis-qgis-0.13.2.tgz: 1f170302cb6db14f2207419a2f29cc81b
[…]
