### 2026-02-18 c7a4c2f
Create ready promise that resolves when model is populated
