### 2026-02-19 be391e7
adding icons (#1129)
