### 2026-02-19 8cb5c50
Simplify
