+++ packages/base/src/widget.ts
@@ -114,31 +114,10 @@ export class JupyterGISPanel extends SplitPanel {
 
     const readyPromise = model.sharedModel.initialSyncReady;
 
-    if (readyPromise) {
-      this._deferredOptions = {
-        model,
-        commandRegistry,
-        formSchemaRegistry,
-        annotationModel,
-      };
-      readyPromise.then(() => this._runDeferredInit());
-    } else {
+    readyPromise.then(() => {
       this._initModel({ model, commandRegistry });
       this._initView(formSchemaRegistry, annotationModel);
-    }
-  }
-
-  private _runDeferredInit(): void {
-    if (this.isDisposed || !this._deferredOptions) {
-      return;
-    }
-    const { model, commandRegistry, formSchemaRegistry, annotationModel } =
-      this._deferredOptions;
-
-    console.log('model.getSelectedStory()', model.getSelectedStory());
-    this._deferredOptions = undefined;
-    this._initModel({ model, commandRegistry });
-    this._initView(formSchemaRegistry, annotationModel);
+    });
   }
 
   _initModel(options: {
@@ -157,9 +136,8 @@ export class JupyterGISPanel extends SplitPanel {
     formSchemaRegistry?: IJGISFormSchemaRegistry,
     annotationModel?: IAnnotationModel,
   ) {
-    const mainViewModel = this._mainViewModel!;
     this._jupyterGISMainViewPanel = new JupyterGISMainViewPanel({
-      mainViewModel,
+      mainViewModel: this._mainViewModel,
       state: this._state,
       formSchemaRegistry: formSchemaRegistry,
       annotationModel: annotationModel,
@@ -170,9 +148,7 @@ export class JupyterGISPanel extends SplitPanel {
 
   get jupyterGISMainViewPanel(): JupyterGISMainViewPanel {
     if (!this._jupyterGISMainViewPanel) {
-      throw new Error(
-        'JupyterGISPanel not ready (initialSyncReady not resolved)',
-      );
+      console.warn('JupyterGISPanel not ready (initialSyncReady not resolved)');
     }
     return this._jupyterGISMainViewPanel;
   }
@@ -182,9 +158,7 @@ export class JupyterGISPanel extends SplitPanel {
     IObservableMap.IChangedArgs<JSONValue>
   > {
     if (!this._view) {
-      throw new Error(
-        'JupyterGISPanel not ready (initialSyncReady not resolved)',
-      );
+      console.warn('JupyterGISPanel not ready (initialSyncReady not resolved)');
     }
     return this._view.changed;
   }
@@ -196,20 +170,17 @@ export class JupyterGISPanel extends SplitPanel {
     if (this.isDisposed) {
       return;
     }
-    this._deferredOptions = undefined;
     if (this._consoleView) {
       this._consoleView.dispose();
     }
     Signal.clearData(this);
-    this._mainViewModel?.dispose();
+    this._mainViewModel.dispose();
     super.dispose();
   }
 
   get currentViewModel(): MainViewModel {
     if (!this._mainViewModel) {
-      throw new Error(
-        'JupyterGISPanel not ready (initialSyncReady not resolved)',
-      );
+      console.warn('JupyterGISPanel not ready (initialSyncReady not resolved)');
     }
     return this._mainViewModel;
   }
@@ -294,15 +265,9 @@ export class JupyterGISPanel extends SplitPanel {
   }
 
   private _state: IStateDB | undefined;
-  private _mainViewModel?: MainViewModel;
-  private _view?: ObservableMap<JSONValue>;
-  private _jupyterGISMainViewPanel?: JupyterGISMainViewPanel;
-  private _deferredOptions?: {
[…]
