+++ packages/schema/src/doc.ts
@@ -25,16 +25,6 @@ import {
   IJupyterGISDocChange,
 } from './interfaces';
 
-const INITIAL_SYNC_KEYS = [
-  'layers',
-  'layerTree',
-  'sources',
-  'stories',
-  'options',
-] as const;
-
-type InitialSyncKey = (typeof INITIAL_SYNC_KEYS)[number];
-
 export class JupyterGISDoc
   extends YDocument<IJupyterGISDocChange>
   implements IJupyterGISDoc
@@ -70,16 +60,12 @@ export class JupyterGISDoc
     return this._initialSyncReadyPromise;
   }
 
-  private _markInitialSyncFired(key: InitialSyncKey): void {
+  private _onOptionsObserverFired(): void {
     if (this._initialSyncResolved) {
       return;
     }
-    this._initialSyncFired.add(key);
-    if (this._initialSyncFired.size === INITIAL_SYNC_KEYS.length) {
-      this._initialSyncResolved = true;
-
-      this._initialSyncReadyResolve();
-    }
+    this._initialSyncResolved = true;
+    this._initialSyncReadyResolve();
   }
 
   getSource(): JSONObject {
@@ -438,13 +424,11 @@ export class JupyterGISDoc
     if (needEmit) {
       this._layersChanged.emit({ layerChange: changes });
     }
-    this._markInitialSyncFired('layers');
   }
 
   private _layerTreeObserver(event: Y.YArrayEvent<IJGISLayerItem>): void {
     const layerTreeChanges = event.delta as Delta<IJGISLayerItem[]>;
     this._layerTreeChanged.emit({ layerTreeChange: layerTreeChanges });
-    this._markInitialSyncFired('layerTree');
   }
 
   private _sourcesObserver(events: Y.YEvent<any>[]): void {
@@ -468,7 +452,6 @@ export class JupyterGISDoc
     if (needEmit) {
       this._sourcesChanged.emit({ sourceChange: changes });
     }
-    this._markInitialSyncFired('sources');
   }
 
   private _storyMapsObserver(events: Y.YEvent<any>[]): void {
@@ -492,7 +475,6 @@ export class JupyterGISDoc
     if (needEmit) {
       this._storyMapsChanged.emit({ storyMapChange: changes });
     }
-    this._markInitialSyncFired('stories');
   }
 
   private _optionsObserver = (event: Y.YMapEvent<Y.Map<string>>): void => {
@@ -505,7 +487,7 @@ export class JupyterGISDoc
       });
     });
     this._optionsChanged.emit(changes);
-    this._markInitialSyncFired('options');
+    this._onOptionsObserverFired();
   };
 
   private _metaObserver = (event: Y.YMapEvent<string>): void => {
@@ -546,6 +528,5 @@ export class JupyterGISDoc
 
   private _initialSyncReadyPromise: Promise<void>;
   private _initialSyncReadyResolve: () => void;
-  private _initialSyncFired = new Set<InitialSyncKey>();
   private _initialSyncResolved = false;
 }
