### 2026-02-20 #9702 @abhishektiwari
Hi @SylvainCorlay, I will be helping to edit this paper. Can you bring word count for `paper.md` to ~1500 – 1800 words? (currently it stands at 3328).

Can you also suggest a 4-5 potential reviewers? You can also share GitHub handles of potential reviewers. Please be sure not to @-tag your suggestions, to avoid unwanted GitHub notifications to potential reviewers. Please also read the our [conflict of interest](https://github.com/openjournals/joss/blob/main/COI.md) (COI) when recommending reviewers.
