+++ ui-tests/tests/notebook.spec.ts
@@ -36,7 +36,7 @@ const testCellOutputs = async (
 
     await page.notebook.runCellByCell({
       onAfterCellRun: async (cellIndex: number) => {
-        await page.waitForTimeout(15000);
+        await page.waitForTimeout(5000);
 
         const cell = await page.notebook.getCellOutputLocator(cellIndex);
         if (cell) {
