+++ packages/schema/src/doc.ts
@@ -25,6 +25,16 @@ import {
   IJupyterGISDocChange,
 } from './interfaces';
 
+/** Default JSON content for a new JupyterGIS document. */
+export const DEFAULT_JGIS_DOCUMENT_CONTENT = `{
+	"schemaVersion": "${SCHEMA_VERSION}",
+	"layers": {},
+	"sources": {},
+	"options": {"latitude": 0, "longitude": 0, "zoom": 0, "bearing": 0, "pitch": 0, "projection": "EPSG:3857"},
+	"layerTree": [],
+	"metadata": {}
+}`;
+
 export class JupyterGISDoc
   extends YDocument<IJupyterGISDocChange>
   implements IJupyterGISDoc
