### 2026-02-20 4d2794a
Use default option values when creating GISDocument
