### 2026-02-20 2534460
Add initialSyncReady promise and defer main view until model is populated (#1130)
* Create ready promise that resolves when model is populated

* Simplify

* Only use options

* test

* Revert

* Unify defaults

* Use default option values when creating GISDocument

* Test test

* Lint

* Update test
