### 2026-02-20 6083164
Publish 0.13.3
SHA256 hashes:

jupytergis-0.13.3-py3-none-any.whl: 3010965368968496947219d50ca1a4927bfb213667187edbd09bc00d813945d2

jupytergis-0.13.3.tar.gz: 49e356420eab0c0d3d0355d7f75568861f44712a3c8d5a2607f3c143aa344b1c

jupytergis-base-0.13.3.tgz: 5c8962db0ed5daaab8808d0b7a426fc67f55dee563a3bc890821c83b9d529afd

jupytergis-jupytergis-core-0.13.3.tgz: 8069864efc7b9c249f7f95747d28cf14743e4c22be303667c1e9d8722ec668f2

jupytergis-jupytergis-lab-0.13.3.tgz: 5f288147d6075dd85d8a7ec75b75b3ec9342f4a8aa45fa8412a00b3c48969dc0

jupytergis-jupytergis-qgis-0.13.3.tgz: d51a803c92203ccdb57bb564ba96a0d1a
[…]
