+++ python/jupytergis_lite/jupytergis/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.13.2"
+__version__ = "0.13.3"
 
 from jupytergis_lab import GISDocument, explore  # noqa
