### 2026-03-29 #9702 @SylvainCorlay
> Hi @SylvainCorlay, I will be helping to edit this paper. Can you bring word count for paper.md to ~1500 – 1800 words? (currently it stands at 3328).
>

@abhishektiwari many thanks for your message. I have shortened the paper as requested.

> Can you also suggest a 4-5 potential reviewers? You can also share GitHub handles of potential reviewers. Please be sure not to @-tag your suggestions, to avoid unwanted GitHub notifications to potential reviewers. Please also read the our [conflict of interest](https://github.com/openjournals/joss/blob/main/COI.md) (COI) when recommending reviewers.
>

Ian Hunt-Isaak ([ianhi](https://github.com/ianhi)) is involved in the Jupyter stack.
Qiusheng Wu  ([giswqs](https://github.com/giswqs)) is listed in your message above is the author of leafmap. Note: he has two commits in the JupyterGIS project.
