### 2026-03-31 #9702 @abhishektiwari
👋 @martinfleis, @TomasBeuzen, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "JupyterGIS: A Collaborative GIS Environment for JupyterLab".

Thanks
Abhi
