### 2026-03-31 #9702 @giswqs
@abhishektiwari I know some of the authors personally, and I have contributed to the JupyterGIS project before. Would that be a conflict of interest for me to review it?
