### 2026-03-31 #9702 @ianhi
Hi @abhishektiwari

I'd be really happy to review this. I've been meaning to check out the software for a while so this is a nice reason to do so.

I read the COI policy and don't feel that I have any, and will be able to make an impartial judgement.  However it bears mention that Martin Renou and have interacted somewhat regularly via gtihub issues over the years as the two most active maintainers of the opensource project https://github.com/matplotlib/ipympl
