### 2026-03-31 #9702 @martinfleis
Hi, this is primarily TypeScript which I do not know. I wouldn't be comfortable reviewing it.
