### 2026-04-05 #9702 @robertmartin8
Hi Abhishek, I’m able to review this as I have past experience with Jupyter and GIS data.
Best,Robert
