### 2026-04-07 #10383 @robertmartin8
Hi @SylvainCorlay, perhaps I'm not looking at the latest proof of the article, but it seems to be missing several of the required sections as per the JOSS standard: the "Software Design" section, "Research Impact Statement", and "AI usage disclosure".

I would suggest condensing the majority of the content in the existing paper (everything after the JupyterGIS heading) into the Software Design section, then adding the aforementioned sections.

The summary and statement of need are very well written, however.
