### 2026-04-07 #10383 @SylvainCorlay
Hi @robertmartin8,

It seems that you are indeed not looking at the latest version of the article, which is available in the `paper` directory of the repository:

https://github.com/geojupyter/jupytergis/blob/main/paper/paper.md

The editorial bot validated the presence of all required sections: https://github.com/openjournals/joss-reviews/issues/9702#issuecomment-4159281742.
