### 2026-04-15 #10383 @gmarupilla
JupyterGIS is right in my wheelhouse — I work with Python geospatial tooling and have experience evaluating research software for usability, docs, and software quality. If you're open to an additional reviewer, I'd be happy to contribute.
