### 2026-04-16 #10383 @robertmartin8
Hi @SylvainCorlay I have completed my initial review.

I assess that JupyterGIS is a mature, functional, and well-documented library and a great contribution to the open source community.  I successfully installed it and confirmed functionality on my machine.

However, I believe the JOSS paper needs to be reworked a little bit (I am now looking at the latest version):

- The paper is **too long** at 1912 words. JOSS papers should be 750-1750 words.
- Generally, I would advocate for a reduction in the use of h2 markdown headings because it can be tough to follow the structure. For example, within State of the Field, you could consider using bullets. The Community and Future Work sections should probably **not** use h2 markdown headers. Installation section should be removed in my opinion.
- The "JupyterGIS" section needs to be significantly compressed: the paper is not supposed to be self-contained documentation, because the JOSS paper links to the repo (from which documentation can be acquired). A "Features" section is fine, but I think it should be condensed.
- While I understand that this is a visual tool, the number of screenshots should be reduced significantly. Perhaps the initial image could be moved below the Summary, because it does help illustrate what JupyterGIS looks like.  I would suggest limiting it to three of the most informative figures.
- There is a small typo within the Software Design section ("Collaborative Aditing").

Best,
Robert
