### 2026-04-20 #10383 @SylvainCorlay
I have shortenned the paper so that it is under the threeshold of 1750 words:
 - removing 4 screenshots,
 - shortening the JupyterGIS section,
 - removing the installation instructions

I have also fixed the typo that was reported.
