### 2026-08-19 #10383 @crvernon
@SylvainCorlay - I just need you to address the following before I accept for publication:

In your paper:

- [ ] LINE 34: “ESRI’s ArcGIS” should be written as “Esri’s ArcGIS”; “Esri” is the proper capitalization of the organization name.
- [ ] LINE 58: Add a comma after “formats”: “…a broad range of geospatial data formats, including vector formats…”
- [ ] LINE 108: The citation rendered as “(Jahns & others, 2015)” is not standard scholarly citation formatting. Correct the bibliography metadata so that the citation renders with the actual author list or an appropriate “et al.” form.
- [ ] LINE 119: “JupyterGIS is integrated in the Galaxy Toolbox” should be written “JupyterGIS is integrated into the Galaxy Toolbox.”
- [ ] LINE 137: “ReadTheDocs” should be written “Read the Docs”, the project’s proper name.
- [ ] LINE 160-163: There is a tense inconsistency: the paragraph begins “We will integrate…” but later says “we ensured.” Change “we ensured” to “we will ensure” or otherwise make the paragraph consistently future tense.
- [ ] LINE 203-207: The author entry “team, J. development.” is malformed. Correct the bibliography metadata so that the Jupyter development team is represented properly.

Let me know when these have been addressed.  Thank you.
