### 2026-08-19 #10383 @SylvainCorlay
Thank you for this new review @crvernon. I have addressed all of your comments and the changes have been integrated into the JupyterGIS repository.
