### 2026-08-19 #10383 @SylvainCorlay
That's great! CI of joss-paper seems to be hanging for now.
